requirements.txt
cmake==3.25.0
fastapi==0.92.0
itsdangerous==2.1.2
keras==2.11.0
numpy==1.24.1
opencv-python==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
retina-face==0.0.13
scikit-image==0.20.0
scipy==1.10.1
uvicorn==0.20.0
Dlib==19.21.1
